+++ pages/editor.tsx
@@ -1,18 +1,19 @@
+import dynamic from 'next/dynamic';
 import React from 'react'
 
-import Head from 'next/head'
-import MDEditor from '@uiw/react-md-editor';
-import '@uiw/react-md-editor/dist/markdown-editor.css'
-import '@uiw/react-markdown-preview/dist/markdown.css';
+const PostEditor = dynamic(
+  () => import('src/components/ToastEditor'),
+  { ssr: false }
+)
 
-export default function Home() {
+//TODO 여기서 이제 API를 이용해서, 받아온 메모 아이디 바탕으로
+//DB에서 메모를 가져오고, 뿌리면 되나...?
+function writePost():JSX.Element {
   return (
-    <div>
-      <Head>
-        <title>Kramo Editor</title>
-        <meta name="description" content="Generated by create next app" />
-      </Head>
-      <MDEditor value='Hello world!'/>
-    </div>
+    <>
+      <PostEditor memo="<br><br><br>제대로 입력되는지 테스트"/>
+    </>
   )
 }
+
+export default writePost;
