+++ pages/ujweditor.tsx
@@ -1,0 +1,18 @@
+import React from 'react'
+
+import Head from 'next/head'
+import MDEditor from '@uiw/react-md-editor';
+import '@uiw/react-md-editor/dist/markdown-editor.css'
+import '@uiw/react-markdown-preview/dist/markdown.css';
+
+export default function Home() {
+  return (
+    <div>
+      <Head>
+        <title>Kramo Editor</title>
+        <meta name="description" content="Generated by create next app" />
+      </Head>
+      <MDEditor value='Hello world!'/>
+    </div>
+  )
+}
